requirements.txt
torch~=2.5.1+cu121
ultralytics~=8.4.9
matplotlib~=3.10.8
numpy~=2.2.6
pillow~=12.1.0